merge to my pouch

+++ ctrd/watch.go
@@ -5,7 +5,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/containerd/containerd"
+	"github.com/alibaba/pouch/pkg/errtypes"
+
 	"github.com/sirupsen/logrus"
 )
 
@@ -16,14 +17,9 @@ type Message struct {
 	err      error
 }
 
-// Error returns the error contained in Message.
-func (m *Message) Error() error {
+// RawError returns the error contained in Message.
+func (m *Message) RawError() error {
 	return m.err
-}
-
-// HasError returns true if the error in message is not nil.
-func (m *Message) HasError() bool {
-	return m.err != nil
 }
 
 // ExitCode returns the exit code in Message.
@@ -38,28 +34,61 @@ func (m *Message) ExitTime() time.Time {
 
 type watch struct {
 	sync.Mutex
-	client     *containerd.Client
-	containers map[string]containerPack
+	containers map[string]*containerPack
 	hooks      []func(string, *Message) error
+
+	// containerdDead to specify whether containerd process is dead
+	containerdDead bool
 }
 
-func (w *watch) add(pack containerPack) {
+func (w *watch) setContainerdDead(isDead bool) error {
 	w.Lock()
 	defer w.Unlock()
 
+	w.containerdDead = isDead
+	return nil
+}
+
+func (w *watch) isContainerdDead() bool {
+	w.Lock()
+	defer w.Unlock()
+
+	return w.containerdDead
+}
+
+func (w *watch) add(pack *containerPack) {
+	w.Lock()
+	defer w.Unlock()
+
+	// TODO(ziren): AcquireQuota may occurred an error
+	// record stream client for grpc client.
+	_ = pack.client.Consume(1)
+
 	w.containers[pack.id] = pack
 
-	go func(pack containerPack) {
+	go func(w *watch, pack *containerPack) {
 		status := <-pack.sch
+
+		// if containerd is dead, the task.Wait channel return is not because
+		// container' task quit, but the channel has broken.
+		// so we just return.
+		if w.isContainerdDead() {
+			return
+		}
 
 		logrus.Infof("the task has quit, id: %s, err: %v, exitcode: %d, time: %v",
 			pack.id, status.Error(), status.ExitCode(), status.ExitTime())
+
+		// Also should release quota when the container destroyed
+		// We should release quota of client that the pack is in using,
+		// not the grpc client executing this parts of code.
+		pack.client.Produce(1)
 
 		if _, err := pack.task.Delete(context.Background()); err != nil {
 			logrus.Errorf("failed to delete task, container id: %s: %v", pack.id, err)
 		}
 
-		if err := pack.container.Delete(context.Background(), containerd.WithSnapshotCleanup); err != nil {
+		if err := pack.container.Delete(context.Background()); err != nil {
 			logrus.Errorf("failed to delete container, container id: %s: %v", pack.id, err)
 		}
 
@@ -69,16 +98,18 @@ func (w *watch) add(pack containerPack) {
 			exitTime: status.ExitTime(),
 		}
 
-		for _, hook := range w.hooks {
-			if err := hook(pack.id, msg); err != nil {
-				logrus.Errorf("failed to execute the exit hooks: %v", err)
-				break
+		if !pack.skipStopHooks {
+			for _, hook := range w.hooks {
+				if err := hook(pack.id, msg); err != nil {
+					logrus.Errorf("failed to execute the exit hooks: %v", err)
+					break
+				}
 			}
 		}
 
 		pack.ch <- msg
 
-	}(pack)
+	}(w, pack)
 
 	logrus.Infof("success to add container, id: %s", pack.id)
 }
@@ -91,13 +122,13 @@ func (w *watch) remove(ctx context.Context, id string) error {
 	return nil
 }
 
-func (w *watch) get(id string) (containerPack, error) {
+func (w *watch) get(id string) (*containerPack, error) {
 	w.Lock()
 	defer w.Unlock()
 
 	pack, ok := w.containers[id]
 	if !ok {
-		return pack, ErrContainerNotfound
+		return pack, errtypes.ErrNotfound
 	}
 	return pack, nil
 }
@@ -110,7 +141,7 @@ func (w *watch) notify(id string) chan *Message {
 	if !ok {
 		ch := make(chan *Message, 1)
 		ch <- &Message{
-			err: ErrContainerNotfound,
+			err: errtypes.ErrNotfound,
 		}
 		return ch
 	}
